+++ src/types.rs
@@ -1,6 +1,8 @@
+mod callback;
 mod client_id;
 mod client_receiver;
 mod client_sender;
 mod subscription_id;
 
-pub(crate) use {client_id::*, client_receiver::*, client_sender::*, subscription_id::*};
+pub use client_id::*;
+pub(crate) use {callback::*, client_receiver::*, client_sender::*, subscription_id::*};